Import PractRand 0.94 and build with CMake
This is a modified PractRand_0.94.zip distribution. The MSVC build configuration
was removed along with prebuilt binaries. This is replaced by a CMake build
script.

Minor changes were made to file paths on disk and in source files. The casing
did not always match, leading to build failures on case-sensitive filesystems.